+++ src/player/damage.js
@@ -3,6 +3,8 @@ import { setInterval } from 'timers/promises'
 
 import { aiter } from 'iterator-helper'
 import combineAsyncIterators from 'combine-async-iterators'
+import Vec3 from 'vec3'
+import mdata from 'minecraft-data'
 
 import { Action, Context } from '../events.js'
 import { create_armor_stand } from '../armor_stand.js'
@@ -10,10 +12,11 @@ import logger from '../logger.js'
 import { GameMode } from '../gamemode.js'
 import { abortable } from '../iterator.js'
 import Entities from '../../data/entities.json' assert { type: 'json' }
-
+import { VERSION } from '../settings.js'
 const DAMAGE_INDICATORS_AMOUNT = 10
 const DAMAGE_INDICATOR_TTL = 1200
 
+const mcData = mdata(VERSION)
 const log = logger(import.meta)
 
 /** @param {import('../context.js').InitialWorld} world */
@@ -24,7 +27,36 @@ export function register({ next_entity_id, ...world }) {
     next_entity_id: next_entity_id + DAMAGE_INDICATORS_AMOUNT,
   }
 }
+/**
+ * Spawns a particle for a player
+ */
+export function world_particle(
+  client,
+  particleName,
+  position,
+  { size = Vec3([1, 1, 1]), count = 1 } = {},
+  blockStateId = null
+) {
+  const particle = mcData.particlesByName[particleName]
 
+  const options = {
+    particleId: particle.id,
+    longDistance: true,
+    ...position,
+    offsetX: size.x,
+    offsetY: size.y,
+    offsetZ: size.z,
+    particleData: 1.0, // maxspeed
+    particles: count,
+    ...(blockStateId != null && {
+      data: {
+        blockState: blockStateId,
+      },
+    }),
+  }
+  client.write('world_particles', options)
+  log.info({ pseudo: client.username, particle }, 'Spawning a particle ')
+}
 export default {
   /** @type {import('../context.js').Reducer} */
   reduce(state, { type, payload }) {
@@ -80,12 +112,18 @@ export default {
             y: y + height - 0.25 + (Math.random() * 2 - 1) * 0.15,
             z: z + (Math.random() * 2 - 1) * 0.25,
           }
-
           if (damage !== undefined) {
             create_armor_stand(client, entity_id, position, {
               text: `-${damage}`,
               color: '#E74C3C', // https://materialui.co/flatuicolors Alizarin
             })
+            world_particle(
+              client,
+              'block',
+              Vec3([x, y + height * 0.7, z]),
+              { count: 10, size: Vec3([0, 0, 0]) },
+              mcData.blocksByName.redstone_block.defaultState
+            )
           } else {
             // Death
             const { xp } = Entities[mob.type]
